# feat: Add performance regression testing to CI pipeline