## Summary

Implements automated performance benchmarking to detect regressions in solver performance across PRs.

Closes #116

## Components Added

### 1. Benchmark Test Suite (`benchmarks/test_solver_performance.py`)
- pytest-benchmark integration for automated testing
- Small and medium problem benchmarks for solvers
- Factory and problem creation overhead benchmarks
- Test markers: `@pytest.mark.slow` for comprehensive tests
- Test groups: `solvers-small`, `solvers-medium`, `factory`, `problem-creation`

**Solver Coverage**:
- ✅ ParticleCollocationSolver (with collocation points)
- ✅ HybridFPParticleHJBFDM (hybrid method)
- ⚠️ FixedPointIterator (requires component solvers - not included yet)

### 2. GitHub Actions Workflow (`.github/workflows/performance_regression.yml`)
- Triggers on PR to `main` with code changes
- Runs benchmarks on both PR and main branches
- Compares results and posts PR comment with formatted report
- Uploads artifacts for 30-day retention
- Configurable regression threshold (15% default)

### 3. Comparison Script (`scripts/compare_benchmarks.py`)
- Analyzes pytest-benchmark JSON results
- Identifies regressions and improvements  
- Generates formatted markdown reports
- Configurable threshold and fail-on-regression option
- CLI tool for local benchmark comparison

## Testing

```bash
# Run all benchmarks
pytest benchmarks/test_solver_performance.py --benchmark-only

# Run fast benchmarks only (exclude @pytest.mark.slow)
pytest benchmarks/test_solver_performance.py -m "not slow" --benchmark-only

# Compare results locally
python scripts/compare_benchmarks.py \
    --pr-results .benchmarks/pr.json \
    --main-results .benchmarks/main.json \
    --threshold 15
```

## CI Integration

- ✅ Runs on every PR modifying solver or benchmark code
- ✅ Posts comparison as PR comment
- ✅ Threshold: 15% regression tolerance
- ✅ Fast execution: 3 rounds minimum for CI speed
- ✅ Artifact uploads for historical tracking

## Benefits

- **Automatic detection** of performance regressions
- **Historical tracking** via CI artifacts
- **Clear visualization** in PR comments
- **Protects recent work** - especially solver unification refactoring

## Example PR Comment

The workflow will post a comment like this:

```
## 📊 Performance Benchmark Results

**Regression threshold**: 15% slower than main

### ✅ No Performance Regressions

All benchmarks are within 15% of main branch performance.

### 📊 Complete Benchmark Comparison

| Benchmark | Main | PR | Change | Status |
|-----------|------|----|---------| --------|
| `test_particle_collocation_small` | 2.45s | 2.41s | -1.6% | ⚪ -1.6% |
| `test_hybrid_solver_small` | 3.12s | 3.15s | +1.0% | ⚪ +1.0% |

---

**Summary**: 2 benchmarks compared, 0 regressions, 0 improvements
```

## Checklist

- [x] Benchmark tests work locally
- [x] GitHub Actions workflow syntax valid
- [x] Comparison script tested  
- [x] Documentation added in code comments
- [x] Pre-commit hooks pass
- [x] Follows CLAUDE.md conventions

## Related

- Implements #116 (performance regression testing)
- Complements #112 (comprehensive benchmarking suite)
- Protects solver unification work (commit 7b3d94c)

🤖 Generated with [Claude Code](https://claude.com/claude-code)